+++ R/tikzDevice-package.R
@@ -104,7 +104,7 @@
 #'
 #'   \item{\code{tikzPdftexWarnUTF}}{ A \code{TRUE/FALSE} value that controls
 #'   whether warnings are printed if Unicode characters are sent to a device
-#'   using the \code{pdftex} engine.  } }
+#'   using the \code{pdftex} engine.  }
 #'
 #'   \item{\code{tikzMetricsMethod}}{ A \code{character} that chooses the
 #'   method used for computing metrics.  This mostly affects the run time for
@@ -128,6 +128,14 @@
 #'         robust method. }
 #'
 #'   Default: \code{"robust"}. }
+#'
+#'   \item{\code{tikzSymbolicColors}}{ A logical value indicating whether colors
+#'   are written as RGB values or as symbolic names in which case the need to be
+#'   defined in the LaTeX document. }
+#'
+#'   \item{\code{tikzMaxSymbolicColors}}{ an integer number indicating the
+#'   maximal number of distinct colors to write symbolically. Any excess color
+#'   will be defined as if \code{symbolicColors} was set to \code{FALSE}.} }
 #'
 #'   Default values for all options may be viewed or restored using the
 #'   \code{\link{setTikzDefaults}} function.